+++ src/server/js/index.js
@@ -10,3 +10,6 @@ import './functions.js';
 import './methods.js';
 import './publish.js';
 import './users_accounts.js';
+
+// on server side, the package is now fully evaluated, so ready
+AccountsManager.ready( true );